apps/coc-discord-bot/cogs/error_monitoring.py
"""
Enhanced Error Monitoring and Logging System for COC-30-Discord-Bot
Provides comprehensive error tracking, analytics, and recovery mechanisms
"""
import asyncio
import traceback
import logging
import psycopg2
import time
from datetime import datetime, timedelta
from typing import Dict, List, Optional, Any
from collections import defaultdict, deque

import discord
from discord.ext import commands, tasks
import psutil

from logging_config import get_logger, log_exception, log_performance
import config
import database_optimized as database

logger = logging.getLogger("error_monitoring")


class ErrorMonitoringCog(commands.Cog):
    def __init__(self, bot):
        self.bot = bot
        # ...existing code...
        # Move all command/event logic here

    # ...existing methods...


async def setup(bot):
    await bot.add_cog(ErrorMonitoringCog(bot))
