apps/coc-discord-bot/requirements.txt
# COC-Discord-Bot Requirements
# Core Discord Bot
discord.py>=2.3.0
discord-py-slash-command>=4.2.1

# Clash of Clans API
aiohttp>=3.8.0
requests>=2.28.0

# Data Processing
tabulate>=0.9.0
python-dateutil>=2.8.0

# Environment Management
python-dotenv>=0.19.0

# Performance Monitoring & System Monitoring
psutil>=5.9.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL driver

# Development Tools (optional)
black>=22.0.0  # Code formatting
flake8>=5.0.0  # Linting
mypy>=0.991    # Type checking

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0
